use common isAtom util
