+++ src/babel/utils.ts
@@ -4,15 +4,29 @@ export function isAtom(
   t: typeof types,
   callee: babel.types.Expression | babel.types.V8IntrinsicIdentifier
 ) {
-  if (t.isIdentifier(callee) && callee.name === 'atom') {
+  if (t.isIdentifier(callee) && atomFunctionNames.includes(callee.name)) {
     return true
   }
 
   if (t.isMemberExpression(callee)) {
     const { property } = callee
-    if (t.isIdentifier(property) && property.name === 'atom') {
+    if (t.isIdentifier(property) && atomFunctionNames.includes(property.name)) {
       return true
     }
   }
   return false
 }
+
+const atomFunctionNames = [
+  'atom',
+  'atomFamily',
+  'atomWithDefault',
+  'atomWithObservable',
+  'atomWithReducer',
+  'atomWithReset',
+  'atomWithStorage',
+  'freezeAtom',
+  'loadable',
+  'selectAtom',
+  'splitAtom',
+]
